Fix duplicate relations on event listeners save (#17408)
* Moved __rawRelationData from Concrete to parent Class

* Removed __rawRelationsData from Concrete class

* Moved __getRawRelationData method to AbstractObject

* Apply php-cs-fixer changes

* Remove classId from Concrete, add to AbstractObject

* Apply php-cs-fixer changes

* Simplify getClassId method

* Update models/DataObject/AbstractObject.php


* Update models/DataObject/AbstractObject.php


---------
